island: keep pairing routing and the home app-grid in the host
Per review, two extractions in this PR pulled UI concerns into island that
don't belong there — island is the headless OEM runtime and must not own host
navigation or UI layout.

- PairingCoordinator (classifyPairFailure + buildPairingRouteStack) hardcoded
  expo-router screen paths and error->screen mapping. The pairing loading and
  success screens own that navigation inline again.
- computeAppGrid was the home-screen tile layout. AppsGrid owns its layout
  useMemo again (restored verbatim; no behavior change).

Keep GlassesReadiness (glasses boot/state) and ConnectionCoordinator (reconnect
policy) — those are device/connection runtime logic, not UI.

Typecheck + jest green (34 suites / 259 tests).
